separate app config, for tests
