+++ src/controllers.ts
@@ -1,4 +1,3 @@
-import { controlerLoader } from './lib/controller/controllerLoader'
 import { KlassController } from './klass/klass.controller'
 import { KlassPrismaRepository } from './klass/klass.prisma'
 import { StudentController } from './students/student.controller'
@@ -6,10 +5,8 @@ import { StudentPrismaRepository } from './students/student.prisma'
 import { SessionController } from './sessions/session.controller'
 import { SessionPrismaRepository } from './sessions/session.prisma'
 
-const controllers = {
+export const controllerConfig = {
   'classes': new KlassController(new KlassPrismaRepository()),
   'students': new StudentController(new StudentPrismaRepository()),
   'sessions': new SessionController(new SessionPrismaRepository()),
 }
-
-export const controllerRouter = controlerLoader(controllers)
