+++ src/server.ts
@@ -1,20 +1,8 @@
-import express from 'express'
-import morgan from 'morgan'
+import { setUpApp } from './app'
 import * as config from './config'
-import * as requestError from './errors'
-import { controllerRouter } from './controllers'
+import { controllerConfig } from './controllers'
 
-const app = express()
-app.use(express.json())
-if(config.env === 'development'){
-  app.use(morgan('dev'))
-}
-
-app.use('/api/v2', controllerRouter)
-
-// Error handling
-app.use(requestError.requestZodErrorHander)
-app.use(requestError.requestErrorHandler)
+const app = setUpApp(controllerConfig)
 
 app.listen(config.PORT, () => {
   console.log(`Express app listen on port: ${config.PORT}`)
